modified:   ainsuForth-CPX_fs-exp.ino 	modified:   ainsuForthsketch.cpp 	modified:   src/kernel/download.cpp 	modified:   src/kernel/getline.cpp 	modified:   src/kernel/getline.h 	modified:   src/kernel/gettoken.cpp 	modified:   src/kernel/gettoken.h
The download word is more advanced, now, and there's a
facility to both detect (only) the '\end.' word, as
well as to print (exactly once) the entire input line
to the console.

Ordinary forth vocabulary is reentrant -- the download
mode is temporary and fits well into a workflow.

Opening a capture file (and printing to it) are next in
line for programming, perhaps.

Thu Aug  3 15:20:17 UTC 2017