modified:   SUCCESS 	modified:   ainsuForth-CPX_fs-exp.ino 	modified:   ainsuForthsketch.cpp 	modified:   src/kernel/download.cpp 	modified:   src/kernel/download.h 	modified:   src/kernel/getline.cpp 	modified:   src/kernel/getline.h 	modified:   src/kernel/gettoken.cpp 	modified:   src/kernel/gettoken.h
Fairly complete implementation of a single-filename solution
to transfer a file to the Adafruit Circuit Playground Express,
from the host PC, via a telecomm program's ASCII upload utility
(can also be hand-typed at the terminal, for testing -- will
test against minicom's ASCII upload soon).

This is called the download word, as it is typed in the forth
interpreter running on the (SAMD21) target.

The CPX has a 2 MB SPI Flash onboard that stores the file.

The file is replayed (as if typing by hand) 'into' the forth
interpreter, where it is either executed, or compiled -- just
as it would be, if typed by hand.  This is known as the 'load'
word, which means 'load a file from SPI flash storage, into
the forth interpreter, where it will be compiled into VM code,
or interpreted (immediately)'.

The 'load' word will also have a silent mode where it does
not echo to console, what it sees (for when you already know
all too well what is in the uploaded source, and do not wish
to see it on console, when it is compiled into VM code, in RAM).

At the moment, it does echo (always).

Thu Aug  3 18:43:18 UTC 2017
4735-b0e-05-