modified:   load.cpp
Use the utterance  '0 echo! load' to load silently (without echo).
Use the utterance '-1 echo! load' to load with echo.

echo in this instance is the usual feedback from the system,
that (often) takes up a lot of space and causes your screen
to scroll a bit much.

A lot of times, what's in the load file is known well enough
to not want to see it replayed on the screen One More Time.

So:

  : false 0 ;  : true -1 ;  false echo! load

Note that compiled and interactive commands are mixed on that line.