modified:   Dictionary.h 	modified:   ainsuForth-CPX_fs-exp.ino 	modified:   flashDict.cpp 	remove_a_file() defined here:   src/kernel/getline.cpp 	modified:   src/kernel/getline.h 	the remove word:   src/kernel/remove.cpp 	modified:   src/kernel/remove.h 	deleted:    tongix
Primitive file management on SPI flashROM.  The new remove word
truncates the existing file, then appends to it a single space
character, and closes the file.  When loaded later, no error
occurs (a single space is tolerated as input, just as it would
be, live at the keyboard, during an interactive session).